try to fix int write

+++ include/stu_printf.h
@@ -11,9 +11,9 @@
 
 int stu_dprintf(int fd, const char *pattern, ...);
 unsigned int stu_strlen(const char *str);
-int nb_len(int str);
+int nb_len(long int str);
 int stu_puts(int fd, const char *str);
-int stu_dputs(int fd, int str);
+int stu_dputs(int fd, long int str);
 int stu_pputs(int fd, unsigned long nbr);
 int stu_pow(int nb, int power);
 char *stu_strdup(const char *str);
